requirements.txt
absl-py==2.0.0
anyio==4.2.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.2
bleach==6.1.0
Bottleneck @ file:///croot/bottleneck_1707864210935/work
Brotli @ file:///croot/brotli-split_1714483155106/work
cachetools==5.3.2
captum==0.7.0
certifi @ file:///croot/certifi_1707229174982/work/certifi
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
cloudpickle==3.0.0
comm==0.2.1
contourpy @ file:///croot/contourpy_1700583582875/work
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
docker-pycreds==0.4.0
executing==2.0.1
fastjsonschema==2.19.1
filelock @ file:///croot/filelock_1700591183607/work
flatbuffers==23.5.26
fonttools==4.44.0
fqdn==1.5.1
gast==0.5.4
gitdb==4.0.10
GitPython==3.1.32
gmpy2 @ file:///work/ci_py311/gmpy2_1676839849213/work
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-pasta==0.2.0
GPUtil==1.4.0
grad-cam==1.5.0
grpcio==1.59.2
h5py==3.10.0
idna @ file:///croot/idna_1714398848350/work
imagecodecs @ file:///croot/imagecodecs_1695064943445/work
imageio @ file:///croot/imageio_1707247282708/work
ipykernel==6.28.0
ipython==8.20.0
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
Jinja2 @ file:///croot/jinja2_1706733616596/work
joblib @ file:///croot/joblib_1713976767370/work
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.4
jupyter_server_terminals==0.5.1
jupyterlab==4.0.10
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
keras==2.15.0
kiwisolver==1.4.5
lazy_loader @ file:///croot/lazy_loader_1695850097191/work
libclang==16.0.6
llvmlite==0.41.1
lpips==0.1.4
Markdown==3.5.1
MarkupSafe @ file:///croot/markupsafe_1704205993651/work
matplotlib==3.8.1
matplotlib-inline==0.1.6
mistune==3.0.2
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
ml-dtypes==0.2.0
mpmath @ file:///croot/mpmath_1690848262763/work
nbclient==0.9.0
nbconvert==7.14.1
nbformat==5.9.2
nest-asyncio==1.5.9
networkx @ file:///croot/networkx_1690561992265/work
notebook==7.0.6
notebook_shim==0.2.3
numba @ file:///croot/numba_1711986258266/work
numexpr @ file:///croot/numexpr_1696515281613/work
numpy==1.26.3
oauthlib==3.2.2
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
opt-einsum==3.3.0
overrides==7.4.0
packaging @ file:///croot/packaging_1710807400464/work
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
patsy==0.5.4
pexpect==4.9.0
pillow @ file:///croot/pillow_1714398848491/work
platformdirs==4.1.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyaml==23.12.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.17.2
pyparsing @ file:///work/ci_py311/pyparsing_1677811559502/work
PySocks @ file:///work/ci_py311/pysocks_1676822712504/work
python-dateutil @ file:///croot/python-dateutil_1715108784345/work
python-json-logger==2.0.7
python-version==0.0.2
pytorch-msssim==1.0.0
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.32.1
requests @ file:///croot/requests_1707355572290/work
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.17.1
rsa==4.9
scikit-image @ file:///croot/scikit-image_1707346116243/work
scikit-learn @ file:///croot/scikit-learn_1714164736727/work
scikit-optimize==0.9.0
scipy==1.10.1
seaborn @ file:///work/ci_py311/seaborn_1676845747354/work
Send2Trash==1.8.2
sentry-sdk==1.29.2
setproctitle==1.3.2
shap @ file:///croot/shap_1692865980380/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
slicer @ file:///tmp/build/80754af9/slicer_1633422823758/work
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
statsmodels==0.14.0
sympy @ file:///croot/sympy_1701397643339/work
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
terminado==0.18.0
threadpoolctl==3.1.0
tifffile @ file:///croot/tifffile_1695107451082/work
tinycss2==1.2.1
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
triton==2.0.0
ttach==0.0.3
types-python-dateutil==2.8.19.20240106
typing_extensions @ file:///croot/typing_extensions_1715268824938/work
tzdata @ file:///croot/python-tzdata_1690578112552/work
unicodedata2 @ file:///croot/unicodedata2_1713212950228/work
uri-template==1.3.0
urllib3 @ file:///croot/urllib3_1707770551213/work
wandb==0.15.8
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
widgetsnbextension==4.0.9
wrapt==1.14.1
